added animations & outsourced intro section
