+++ src/app/pages/landing-page/sections/tutorial/tutorial.component.html
@@ -2,7 +2,7 @@
   <div class="container">
     <div class="row">
       <div class="col-md-6">
-        <div class="tutorial_section_left">
+        <div class="tutorial_section_left a_lightSpeedInLeft">
           <div class="sec-heading">
             <div class="sub-inner mb-15">
               <span class="sub-title">WIE RICHTE ICH DEN DISCORD-BOT EIN?</span>
@@ -64,7 +64,7 @@
 
       <div class="col-md-6">
         <div class="tutorial_section_right_content">
-          <div class="tutorial_section_card tutorial_section_card1 !items-center">
+          <div class="tutorial_section_card a_lightSpeedInRight">
             <div class="tutorial_section_card_number">
               <img ngSrc="assets/img/icons/steps/one.png" alt="Clank ~ Discord-Bot (Step One)"
                    class="img-fluid" height="100" width="100" />
@@ -75,7 +75,7 @@
                 und gib ihm alle Rechte, nach denen er dich bei der Einladung fragt.</p>
             </div>
           </div>
-          <div class="tutorial_section_card tutorial_section_card2 !items-center">
+          <div class="tutorial_section_card a_lightSpeedInRight animate__delay-1s">
             <div class="tutorial_section_card_number">
               <img ngSrc="assets/img/icons/steps/two.png" alt="Clank ~ Discord-Bot (Step Two)"
                    class="img-fluid" height="100" width="100" />
@@ -86,7 +86,7 @@
                 <code>/bot-sprache</code> (oder <code>/bot-language</code>) verwendest.</p>
             </div>
           </div>
-          <div class="tutorial_section_card tutorial_section_card3 !items-center">
+          <div class="tutorial_section_card a_lightSpeedInRight animate__delay-2s">
             <div class="tutorial_section_card_number">
               <img ngSrc="assets/img/icons/steps/three.png" alt="Clank ~ Discord-Bot (Step Three)"
                    class="img-fluid" height="100" width="100" />
