+++ src/app/pages/landing-page/sections/tutorial/tutorial.component.scss
@@ -32,7 +32,7 @@
 
 // semi-transparent background for tutorial step cards
 .tutorial_section_card {
-  @apply flex items-start justify-start min-h-48 py-9 px-12 backdrop-blur-[10px] mb-5;
+  @apply flex items-center justify-start min-h-48 py-9 px-12 backdrop-blur-[10px] mb-5;
   background: rgba(30, 31, 53, 0.7);
 }
