+++ src/app/pages/landing-page/sections/tutorial/tutorial.component.spec.ts
@@ -1,6 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { TutorialComponent } from './tutorial.component';
+import {ActivatedRoute} from "@angular/router";
 
 describe('TutorialComponent', () => {
   let component: TutorialComponent;
@@ -8,10 +9,11 @@ describe('TutorialComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [TutorialComponent]
+      imports: [TutorialComponent],
+      providers: [ { provide: ActivatedRoute, useValue: {} }]
     })
     .compileComponents();
-    
+
     fixture = TestBed.createComponent(TutorialComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
